requirements.txt
streamlit
numpy
pandas
scikit-learn
gensim
matplotlib
jieba
wordcloud
python-docx
PyPDF2
requests
openai
zhipuai
torch
Pillow
